requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.2
billiard==4.2.2
Brotli==1.1.0
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
convertdate==2.4.0
coverage==7.11.3
cron_descriptor==2.0.6
cssselect2==0.8.0
diff-match-patch==20241021
Django==5.2.8
django-celery-beat==2.8.1
django-crispy-forms==2.5
django-environ==0.12.0
django-import-export==4.3.14
django-leaflet==0.32.0
django-ninja==1.5.0
django-simple-history==3.10.1
django-solo==2.4.0
django-timezone-field==7.1
django-unfold==1.0.0
execnet==2.1.2
factory_boy==3.3.3
Faker==37.12.0
fonttools==4.60.1
geographiclib==2.1
geopy==2.4.1
gunicorn==23.0.0
idna==3.10
iniconfig==2.3.0
kombu==5.5.4
lunardate==0.2.2
packaging==25.0
pi==0.1.2
pillow==12.0.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydyf==0.11.0
Pygments==2.19.2
pyluach==2.3.0
PyMeeus==0.5.12
pyphen==0.17.2
pytest==9.0.1
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-xdist==3.8.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
redis==6.4.0
reportlab==4.4.4
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.11.0
workalendar==17.0.0
zopfli==0.4.0
